captioner_functionApp/requirements.txt
#Ignore this if running locally, this is only here for deployment purposes
azure-functions
azure-storage-blob
Pillow
moondream
requests